+++ generate_vectors.py
@@ -30,7 +30,7 @@ def decode_mnemonic(mnemonic):
 def generate_mnemonics_random(group_threshold, groups):
     secret = random_bytes(16)
     return shamir.generate_mnemonics(
-        group_threshold, groups, secret, iteration_exponent=0
+        group_threshold, groups, secret, extendable=False, iteration_exponent=0
     )
 
 
@@ -46,7 +46,7 @@ if __name__ == "__main__":
         description = "Valid mnemonic without sharing ({} bits)"
         secret = random_bytes(n)
         groups = shamir.generate_mnemonics(
-            1, [(1, 1)], secret, b"TREZOR", iteration_exponent=0
+            1, [(1, 1)], secret, b"TREZOR", extendable=False, iteration_exponent=0
         )
         output(description.format(8 * n), groups[0], secret)
 
@@ -63,13 +63,16 @@ if __name__ == "__main__":
             indices[4] += 1 << overflowing_bits
             indices = indices[: -constants.CHECKSUM_LENGTH_WORDS]
             mnemonic = wordlist.mnemonic_from_indices(
-                indices + rs1024.create_checksum(indices)
+                indices
+                + rs1024.create_checksum(indices, constants.CUSTOMIZATION_STRING_ORIG)
             )
             output(description.format(8 * n), [mnemonic], b"")
 
         description = "Basic sharing 2-of-3 ({} bits)"
         secret = random_bytes(n)
-        groups = shamir.generate_mnemonics(1, [(2, 3)], secret, b"TREZOR", 2)
+        groups = shamir.generate_mnemonics(
+            1, [(2, 3)], secret, b"TREZOR", extendable=False, iteration_exponent=2
+        )
         output(description.format(8 * n), random.sample(groups[0], 2), secret)
         output(description.format(8 * n), random.sample(groups[0], 1), b"")
 
@@ -83,21 +86,21 @@ if __name__ == "__main__":
         description = "Mnemonics with different iteration exponents ({} bits)"
         groups = generate_mnemonics_random(1, [(2, 2)])
         data = decode_mnemonic(groups[0][0])
-        data[1] = 3  # change iteration exponent from 0 to 3
+        data[2] = 3  # change iteration exponent from 0 to 3
         mnemonics = [encode_mnemonic(*data), groups[0][1]]
         output(description.format(8 * n), mnemonics, b"")
 
         description = "Mnemonics with mismatching group thresholds ({} bits)"
         groups = generate_mnemonics_random(2, [(1, 1), (2, 2)])
         data = decode_mnemonic(groups[0][0])
-        data[3] = 1  # change group threshold from 2 to 1
+        data[4] = 1  # change group threshold from 2 to 1
         mnemonics = groups[1] + [encode_mnemonic(*data)]
         output(description.format(8 * n), mnemonics, b"")
 
         description = "Mnemonics with mismatching group counts ({} bits)"
         groups = generate_mnemonics_random(1, [(2, 2)])
         data = decode_mnemonic(groups[0][0])
-        data[4] = 3  # change group count from 1 to 3
+        data[5] = 3  # change group count from 1 to 3
         mnemonics = [encode_mnemonic(*data), groups[0][1]]
         output(description.format(8 * n), mnemonics, b"")
 
@@ -109,35 +112,40 @@ if __name__ == "__main__":
         for group in groups:
             for mnemonic in group:
                 data = decode_mnemonic(mnemonic)
-                data[4] = 1  # change group count from 2 to 1
+                data[5] = 1  # change group count from 2 to 1
                 mnemonics.append(encode_mnemonic(*data))
         output(description.format(8 * n), mnemonics, b"")
 
         description = "Mnemonics with duplicate member indices ({} bits)"
         groups = generate_mnemonics_random(1, [(2, 3)])
         data = decode_mnemonic(groups[0][0])
-        data[5] = 2  # change member index from 0 to 2
+        data[6] = 2  # change member index from 0 to 2
         mnemonics = [encode_mnemonic(*data), groups[0][2]]
         output(description.format(8 * n), mnemonics, b"")
 
         description = "Mnemonics with mismatching member thresholds ({} bits)"
         groups = generate_mnemonics_random(1, [(2, 2)])
         data = decode_mnemonic(groups[0][0])
-        data[6] = 1  # change member threshold from 2 to 1
+        data[7] = 1  # change member threshold from 2 to 1
         mnemonics = [encode_mnemonic(*data), groups[0][1]]
         output(description.format(8 * n), mnemonics, b"")
 
         description = "Mnemonics giving an invalid digest ({} bits)"
         groups = generate_mnemonics_random(1, [(2, 2)])
         data = decode_mnemonic(groups[0][0])
-        data[7] = bytes((data[7][0] ^ 1,)) + data[7][1:]  # modify the share value
+        data[8] = bytes((data[8][0] ^ 1,)) + data[8][1:]  # modify the share value
         mnemonics = [encode_mnemonic(*data), groups[0][1]]
         output(description.format(8 * n), mnemonics, b"")
 
         # Group sharing.
         secret = random_bytes(n)
         groups = shamir.generate_mnemonics(
-            2, [(1, 1), (1, 1), (3, 5), (2, 6)], secret, b"TREZOR", iteration_exponent=0
+            2,
+            [(1, 1), (1, 1), (3, 5), (2, 6)],
+            secret,
+            b"TREZOR",
+            extendable=False,
+            iteration_exponent=0,
         )
 
         description = "Insufficient number of groups ({} bits, case {})"
@@ -163,14 +171,38 @@ if __name__ == "__main__":
     description = "Mnemonic with insufficient length"
     secret = random_bytes((shamir.MIN_STRENGTH_BITS // 8) - 2)
     identifier = random.randrange(1 << shamir.ID_LENGTH_BITS)
-    mnemonic = encode_mnemonic(identifier, 0, 0, 1, 1, 0, 1, secret)
+    mnemonic = encode_mnemonic(identifier, False, 0, 0, 1, 1, 0, 1, secret)
     output(description, [mnemonic], b"")
 
     description = "Mnemonic with invalid master secret length"
     secret = b"\xff" + random_bytes(shamir.MIN_STRENGTH_BITS // 8)
     identifier = random.randrange(1 << shamir.ID_LENGTH_BITS)
-    mnemonic = encode_mnemonic(identifier, 0, 0, 1, 1, 0, 1, secret)
+    mnemonic = encode_mnemonic(identifier, False, 0, 0, 1, 1, 0, 1, secret)
     output(description, [mnemonic], b"")
+
+    description = "Valid mnemonics which can detect some errors in modular arithmetic"
+    secret = b"\xado*\xd8\xb5\x9b\xbb\xaa\x016\x9b\x90\x06 \x8d\x9a"
+    mnemonics = [
+        "herald flea academic cage avoid space trend estate dryer hairy evoke eyebrow improve airline artwork garlic premium duration prevent oven",
+        "herald flea academic client blue skunk class goat luxury deny presence impulse graduate clay join blanket bulge survive dish necklace",
+        "herald flea academic acne advance fused brother frozen broken game ranked ajar already believe check install theory angry exercise adult",
+    ]
+    output(description, mnemonics, secret)
+
+    for n in [16, 32]:
+        description = "Valid extendable mnemonic without sharing ({} bits)"
+        secret = random_bytes(n)
+        groups = shamir.generate_mnemonics(
+            1, [(1, 1)], secret, b"TREZOR", extendable=True, iteration_exponent=3
+        )
+        output(description.format(8 * n), groups[0], secret)
+
+        description = "Extendable basic sharing 2-of-3 ({} bits)"
+        secret = random_bytes(n)
+        groups = shamir.generate_mnemonics(
+            1, [(2, 3)], secret, b"TREZOR", extendable=True, iteration_exponent=0
+        )
+        output(description.format(8 * n), random.sample(groups[0], 2), secret)
 
     with open("vectors.json", "w") as f:
         json.dump(